Introduce additional protection methond.
We introduce a way to protect inside a region when
we not directly in that region. This may be needed
when we run internal computation that is in the different
monad (ST or IO for example) and we do not have a way to proove
(or carry region to that monad). For this purpose we
introduce MonadRHandler, that carry a pointer to MonadR state
and allocHandler method that creates such handler. MonadRHandler
is quantified by MonadR Region variable, as a result MonadRHandler
can't escape outside of the MonadR scope.